feat: Restore ActionCard component extraction without missing lines
